chore: 🔧 move Calendar module into MeetingSetup
